refactor : useFocus
